12.1.Adding & Checking the User's Password{Adding the password Field to User}
Until now, we've allowed users to login without any password. As much fun as it would be to deploy this to production... I think we should probably fix that. If you look at your User class, our users actually don't have a password field at all.

When you originally use the make:user command, you can tell it to create this field for you. We told it to not do this... just to keep things simpler as we were learning. So, we'll do it now.

Adding the password Field to User
Find your terminal and run:

php bin/console make:entity
Update the User class to add a new field called password. Make it a string with length 255. It doesn't need to be quite that long, but that's fine. Can it be null? Say no: in our system, each user will always have a password.
And... done! It updated the User.php file, but it did not generate the normal getPassword() method because we already had that method before. We'll check that out in a minute.
Before that, run:

php bin/console make:migration
Move over and check out the Migrations directory. Open the new file and... yes! It looks perfect: ALTER TABLE user ADD password
Close that, go back to your terminal, and migrate:

php bin/console doctrine:migrations:migrate
Awesome!