12.2.Adding & Checking the User's Password{Updating the User Class}
Updating the User Class
Go open the User class. Yep - we now have a password field. And all the way at the bottom, a setPassword() method.
Scroll up to find getPassword(). This already existed from back when our user had no password. Now that it does, return $this->password. Oh, and just to be clear, this password will not be a plain-text password. No, no, no! The string that we store in the database will always be properly salted & encoded. In fact, look at the method below this: getSalt().
In reality, there are two things you need to store in the database: the encoded password and the random salt value that was used to encode the password.
But, great news! Most modern encoders - including the one we will use - store the salt value as part of the encoded password string. In other words, we only need this one field. And, the getSalt() method can stay blank. I'll update the comment to explain why.
I love doing no work!