13.2.access_control Authorization & Roles{Roles! }
Roles!
Let's talk about how roles work in Symfony: it's simple and it's beautiful. Down on the web debug toolbar, click on the user icon. Cool: we're logged in as spacebar1@example.com and we have one role: ROLE_USER. Here's the idea: when a user logs in, you give them whatever "roles" you want - like ROLE_USER. Then, you run around your code and make different URLs require different roles. Because our user does not have ROLE_ADMIN, we are denied access.
But... why does our user have ROLE_USER? I don't remember doing anything with roles during the login code. Open the User class. When we ran the make:user command, one of the methods that it generated was getRoles().
Look at it carefully: it reads a roles property, which is an array that's stored in the database.
Right now, this property is empty for every user in the database: we have not set this to any value in the fixtures.
But, inside getRoles(), there's a little extra logic that guarantees that every user at least has this one role: ROLE_USER.
This is nice because we now know that, if you are logged in, you definitely have this one role. Also... you need to make sure that getRoles() always returns at least one role... otherwise weird stuff happens: the user becomes an undead zombie that is "sort of" logged in.
To prove that this roles system works like we expect, change ROLE_ADMIN to ROLE_USER in the access control.
Then, click back to the admin page and... access granted!