14.2.Target Path: Redirecting an Anonymous User{Redirecting Back on Success)
Hmm: how could we update this method to send the user back to the previous page instead?
Symfony can help with this. Find your browser, log out, and then go back to /admin/comment. Whenever you try to access a URL as an anonymous user, before Symfony redirects to the login page, it saves this URL - /admin/comment - into the session on a special key. So, if we can read that value from the session inside onAuthenticationSuccess(), we can redirect the user back there!
To do this, at the top of your authenticator, use a trait TargetPathTrait.
Then, down in onAuthenticationSuccess(), add if $targetPath = $this->getTargetPath(). This method comes from our handy trait! It needs the session - $request->getSession() - and the "provider key", which is actually an argument to this method.
The provider key is just the name of your firewall... but that's not too important here.
Oh, and, yea, the if statement might look funny to you: I'm assigning the $targetPath variable and then checking to see if it's empty or not. If it's not empty, if there is something stored in the session, return new RedirectResponse($targetPath).
That's it! If there is no target path in the session - which can happen if the user went to the login page directly - fallback to the homepage.
Let's try it! Log back in... with password engage. Yea! Got it! I know, it feels weird to celebrate when you see an access denied page. But we expected that part. The important thing is that we were redirected back to the page we originally tried to access. That's excellent UX.
Next - as nice as access controls are, we need more granular control. Let's learn how to control user access from inside the controller.