16.1.Dynamic Roles{php bin/console make:controller}
I want to create something new: a new user account page. Find your terminal and run:

php bin/console make:controller
Create a new AccountController. Open that up
Perfect! A new /account page, which we can see instantly if we go to that URL.
Change the route name to app_account to be consistent with our other code. And, I'm not going to pass any variables to the template for now. Open that: templates/account/index.html.twig. Let's customize this just a bit: Manage Account and an h1: Manage your Account. That's pretty boring... but it should be enough for us to get into trouble!
Ok: I only want this page to be accessible by users who are logged in. Log out and then go back to /account. Obviously, right now, anybody can access this. Hmm: we do already know how to require the user to have a specific role to access something - like in CommentAdminController where we require ROLE_ADMIN