16.2.Dynamic Roles{Check if the User Is Logged In}
Check if the User Is Logged In
But... how can we make sure that the user is simply... logged in?
There are actually two different ways. I'll tell you about one of those ways later. But right now, I want to tell you about the easier of the two ways: just check for ROLE_USER.
Above the AccountController class - so that it applies to any future methods - add @IsGranted("ROLE_USER").
So... why is this a valid way to check that the user is simply logged in? Because, remember! In User, our getRoles() method is written so that every user always has at least this role.
If you are logged in, you definitely have ROLE_USER.
Refresh the page now: it bumps us to the login page. Log in with password engage and... nice! We're sent back over to /account. Smooth.