16.4.Dynamic Roles{Checking for Roles in Twig}
Checking for Roles in Twig
Oh, and now that we know how to check if the user is logged in, let's fix our user drop-down: we should not show the login link once we're logged in.
In PhpStorm, open templates/base.html.twig and scroll down a bit. Earlier, when we added the login link, we commented out our big user drop-down.
Now, we can be a bit smarter. Copy that entire section: we will show it when the user is logged in.
Oh, but how can we check if the user has a role from inside Twig? With: is_granted(): if is_granted('ROLE_USER'), else - I'll indent my logout link - and endif. Inside the if, paste the drop-down code.
Ah! Let's go see it! Refresh! Our user drop-down is back! Oh, except all of these links go... nowhere. We can fix that!
For profile, that route is app_account: path('app_account'). For logout, that's path('app_logout').
And, for "Create Post", we haven't built that yet. But, there is a controller called ArticleAdminController and we have at least started this. Give this route a name="admin_article_new".
 Tip

Oh! And don't forget to require ROLE_ADMIN on the controller!
We'll link here, even though it's not done. Oh, but this link is only for admin users. So, surround this with is_granted("ROLE_ADMIN").
Nice! Let's make sure we didn't mess up - refresh! Woohoo! Because we are logged in as an admin user, we see the user drop-down and the Create Post link.
Next: we need to talk about a few unique roles that start with IS_AUTHENTICATED and how these can be used in access_control to easily require login for every page on your site.