17.1.IS_AUTHENTICATED_ & Protecting All URLs{IS_AUTHENTICATED_FULLY; Web Debug Toolbar & Access Control Checks}
I mentioned earlier that there are two ways to check whether or not the user is simply logged in. The first is by checking ROLE_USER. I like this one because it's simple. It works because of how our getRoles() method is written.
The only reason I'm even going to mention the second way is because I want you to know what it is if you see it, and, it leads us towards a few other interesting things.

IS_AUTHENTICATED_FULLY
Let's play a little bit in security.yaml. Under access_control add a new entry with path ^/account. Yes, this will be a totally redundant access control because we're already requiring ROLE_USER from inside the controller.
Just pretend that we don't have this controller code for a minute.
On your access_control, if you wanted to require the user to be logged in, you could use roles: ROLE_USER or IS_AUTHENTICATED_FULLY
OoooOOOoo.
Well, it's not really that fancy: it's just a special string that simply checks if the user is logged in or not. In our system, it's 100% identical to ROLE_USER.
Move over, go back to /account and... yep! Access is still granted.

Web Debug Toolbar & Access Control Checks
Oh, and I want to show you something cool! Click the little security icon on the web debug toolbar. This has some pretty sweet stuff in it. In addition to saying who you're logged in as and your roles, it also has a table down here with some lower-level info. But what I really want to show you is all the way at the bottom. Yes! The access decision log. This records every time that we checked whether or not the user had access to something on this page. The first check is for IS_AUTHENTICATED_FULLY from access_control. Granted! Then, two ROLE_USER checks and one ROLE_ADMIN check.

One of those ROLE_USER checks is from AccountController:

 24 lines  src/Controller/AccountController.php
/**
 * @IsGranted("ROLE_USER")
 */
class AccountController extends AbstractController
{
}
And the other comes from is_granted() in the template. The ROLE_ADMIN check also lives here:

 77 lines  templates/base.html.twig
<html lang="en">
    <body>
        <nav class="navbar navbar-expand-lg navbar-dark navbar-bg mb-5">
            <div class="collapse navbar-collapse" id="navbarNavDropdown">
                <ul class="navbar-nav ml-auto">
                    {% if is_granted('ROLE_USER') %}
                        <li class="nav-item dropdown" style="margin-right: 75px;">
                            <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
                                {% if is_granted('ROLE_ADMIN') %}
                                {% endif %}
                            </div>
                        </li>
                    {% endif %}
                </ul>
            </div>
        </nav>
    </body>
</html>
So, this is just a nice way to debug all the security checks happening on your page.