17.2.IS_AUTHENTICATED_ & Protecting All URLs{Requiring Login on Every Page}
Requiring Login on Every Page
Anyways, we now know IS_AUTHENTICATED_FULLY is a way to check if the user is logged in. Though... because of the way our app is written, checking ROLE_USER does the same thing and... it's shorter to write.
But! This does touch on another interesting topic. This is a news site, so most of the pages will be accessible to anonymous users. We'll require login on just the pages that need it. Not all sites are like this, however. On some sites, you want to do the opposite: you want to require authentication for every page, or at least, almost every page. In those cases, a better strategy is to require login on all pages and then allow anonymous access on just a few pages.
We can do this by being clever with access_control. Try this: change the path to just ^/.
Because this is a regular expression, it will match every URL and so every page now requires login.
If we refresh, we still have access. But now, log out!
Whoa! The page is broken! Like, crazy broken! localhost redirected too many times? Yep, our security system is too awesome. Because we're now anonymous, when we try to access any page, we're redirected to /login. But guess what? /login requires authentication too! So what does Symfony do? It redirects us to /login!