18.1.Fetch the User Object
Ok, because our site will be mostly public, I'll comment-out these examples.
Next, let's learn how to find out who is logged in by fetching their User object. Or two, figure out who is logged in.
That's exactly what we need to do in AccountController so that we can start printing out details about the user's account. So... how can we find out who is logged in? With $this->getUser().
Go back to your browser and head to /account. Nice! This gives us the User entity object! That's awesome because we can do all kinds of cool stuff with it. For example, let's see if we can log the email address of who is logged in.