19.3.Custom User Method{bin/console doctrine:query:sql "SELECT * FROM user"}
Finally! Let's get to work in account/index.html.twig. Replace the ? marks with app.user.twitterUsername.
Hmm, but we probably don't want to show this block if they don't have a twitterUsername. No problem: surround this with an if statement.
Perfect! Ok, let's go find a user that has their twitterUsername set! Run:

php bin/console doctrine:query:sql "SELECT * FROM user"
Scroll up and... cool: spacebar1@example.com. Move back to your browser and refresh. Oh! We got logged out! That's because the id of the user that we were logged in as was removed from the database when we reloaded the fixtures.
Login as spacebar1@example.com, password engage. Click sign in and... got it!