19.4.Custom User Method{Custom User Method for RoboHash}
Custom User Method for RoboHash
Oh, and there's one other thing that we can finally update! See the user avatar on the drop-down? That's totally hardcoded. Let's roboticize that! Yea, roboticize apparently is a real word.
Copy the src for the RoboHash. Then, open up base.html.twig and, instead of pointing to the astronaut's profile image, paste it!
Try it! Move over and... refresh!
Nice! But, hmm... there is one small thing that I don't like. Right click on the image, copy the image address and paste in a new tab. Oh. That's a pretty big image: 300x300. It's not a huge deal, but our users are downloading a pretty big image, just to display this teenie-tiny thumbnail.
Fortunately, the fine people who created RoboHash added a feature to help us! By adding ?s=100x100, we can get a smaller image. Let's do that on the menu.
But, wait! Instead of just putting ?size= right here... let's get organized! I don't like duplicating the RoboHash link everywhere. Open your User class. Let's add a new custom function called public function getAvatarUrl().
We don't actually have an avatarUrl property... but that's ok! Give this an int argument that's optional and the method will return a string. Inside, set $url = and paste the RoboHash link. Remove the email but add $this->getEmail(). Easy enough! For the size part, if a  is passed in, use $url .= to add sprintf('?size=%dx%d'), passing  for both of these wildcards. At the bottom, return .
Now that we're done with our fancy new function, go into index.html.twig, remove the long string, and just print app.user.avatarUrl.
We can reference avatarUrl like a property, but behind the scenes, we know that Twig is smart enough to call the getAvatarUrl() method.
Copy that, go back into base.html.twig and paste. But this time, call it like a function: pass 100.
Let's see if it works! Close a tab then, refresh! Yep! And if we copy the image address again and load it... nice! A little bit smaller.
Next, let's find out how to fetch the user object from the one spot we haven't talked about yet: services.