22.3.Impersonation (switch_user){Adding a Banner when you are Impersonating; Adding Query Parameters with path()}
Adding a Banner when you are Impersonating
Anyways, to exit and return to your normal identity, find a phone booth, close the door, and add ?_switch_user=_exit to any URL. And... we're back to being us!
Switch one more time back to spacebar1@example.com. One of the only issues with _switch_user is that it's not super obvious that we're switched! Yep, you might switch to a user, go check Facebook, then come back, forget that you're still switched to them, and start commenting on their behalf. What? No, I've definitely never done this... I'm just saying it's possible.
To prevent these... awkward situations, let's put a big banner on top when we're switched. Open base.html.twig and find the body tag. Here's the key: when we are switched to another user, Symfony gives us a special role called ROLE_PREVIOUS_ADMIN. We can use that to our advantage: if is_granted('ROLE_PREVIOUS_ADMIN'), then print an alert block. Inside, say:
  You are currently switched to this user
And, to maximize our fanciness, let's add a link to exit. Use the path function to point to app_homepage. For the second argument, pass an array with the necessary _switch_user set to _exit. At the end, say "Exit Impersonation"

Adding Query Parameters with path()
Let's see how it looks! Move over and refresh! Nice! Even I won't forget when I'm impersonating. And, check out the URL on the link: it's perfect - ?_switch_user=_exit. But... wait... the way we just used the path() function was a bit weird.
Why? Open templates/article/homepage.html.twig and find the article list. You might remember that the second argument of the path() function is normally used to fill in the "wild card" values for a route:
 65 lines  templates/article/homepage.html.twig
{% block body %}
    <div class="container">
        <div class="row">
            <div class="col-sm-12 col-md-8">
                <div class="article-container my-1">
                    <a href="{{ path('article_show', {slug: article.slug}) }}">
                </div>
            </div>
        </div>
    </div>
{% endblock %}
Hold Command or Control and click article_show. Yep! This route has a {slug} wild card

 65 lines  src/Controller/ArticleController.php
class ArticleController extends AbstractController
{
    /**
     * @route("/news/{slug}", name="article_show")
     */
    public function show(Article $article)
    {
    }
}
And so, when we link to it, we need to pass a value for that slug wildcard via the 2nd argument to path().
We already knew that. And this is the normal purpose of the second argument to path(). However, if you pass a key to the second argument, and that route does not have a wildcard with that name, Symfony just adds it as a query parameter.
That is why we can click this link to exit impersonation.
Next - let's build an API endpoint with Symfony's serializer! That will be our first step towards API authentication.