23.1.Serializer & API Endpoint{Creating the API Endpoint}
In addition to our login form authentication, I also want to allow users to log in by sending an API token. But, before we get there, let's make a proper API endpoint first.

Creating the API Endpoint
I'll close a few files and open AccountController. To keep things simple, we'll create an API endpoint right here. Add a public function at the bottom called accountApi().
This new endpoint will return the JSON representation of whoever is logged in. Above, add @route("/api/account") with name="api_account".
The code here is simple - excitingly simple! $user = $this->getUser() to find who's logged in.
We can safely do this thanks to the annotation on the class: every method requires authentication. Then, to transform the User object into JSON - this is pretty cool - return $this->json() and pass $user.
Let's try it! In your browser, head over to /api/account. And! Oh! That's not what I expected! It's JSON... but it's totally empty!