23.3.Serializer & API Endpoint{Serialization Groups}
Serialization Groups
Except... well... we probably don't want to include all of these properties - especially the encoded password. I know, I said we weren't going to talk about the serializer, and yet, I do want to fix this one thing!
Open your User class. To control which fields are serialized, above each property, you can use an annotation to organize into "groups". I won't expose the id, but let's expose email by putting it into a group: @groups("main"). When I auto-completed that annotation, the PHP Annotations plugin added the use statement I need to the top of the file.
Oh, and I totally invented the "main" part - that's the group name, and you'll see how I use it in a minute. Copy the annotation and also add firstName and twitterUsername to that same group.
To complete this, in AccountController, we just need to tell the json() method to only serialize properties that are in the group called "main". To do that, pass the normal 200 status code as the second argument, we don't need any custom headers, but we do want to pass one item to "context". Set groups => an array with the string main.
You can include just one group name here like this, or tell the serializer to serialize the properties from multiple groups.
Let's try it! Refresh! Yes! Just these three fields.
Ok, we are now ready to take on a big, cool topic: API token authentication.